Replace int64 id type with uint64

+++ kubeflake.go
@@ -27,14 +27,14 @@ func machineID() (uint16, error) {
 	return uint16(machineID), nil
 }
 
-func New() (int64, error) {
+func New() (uint64, error) {
 	id, err := sf.NextID()
 	if err != nil {
 		return 0, err
 	}
-	return int64(id), nil
+	return id, nil
 }
 
-func Must(id int64, _ error) int64 {
+func Must(id uint64, _ error) uint64 {
 	return id
 }
